pyproject.toml
[tool.poetry]
name = "event-management-system-backend"
version = "0.1.0"
description = ""
authors = ["Akbota <aqbotamaratqyzy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
celery = "^5.4.0"
redis = "^5.0.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
django-celery-beat = "^2.6.0"
python-decouple = "^3.8"
django-celery-results = "^2.5.1"
tools = "^0.1.9"
redis-tools = "^0.0.2"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
